refactor(console): move llm-connect component to agents page directory
